move timezone tests where they belong
