+++ hibernate-core/src/test/java/org/hibernate/orm/test/timezones/DefaultZonedTest.java
@@ -1,15 +1,13 @@
-package org.hibernate.userguide.mapping.basic;
+package org.hibernate.orm.test.timezones;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
-import org.hibernate.cfg.AvailableSettings;
 import org.hibernate.dialect.SybaseDialect;
+import org.hibernate.dialect.TimeZoneSupport;
 import org.hibernate.testing.orm.junit.DomainModel;
-import org.hibernate.testing.orm.junit.ServiceRegistry;
 import org.hibernate.testing.orm.junit.SessionFactory;
 import org.hibernate.testing.orm.junit.SessionFactoryScope;
-import org.hibernate.testing.orm.junit.Setting;
 import org.junit.jupiter.api.Test;
 
 import java.time.OffsetDateTime;
@@ -19,10 +17,9 @@ import java.time.ZonedDateTime;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@DomainModel(annotatedClasses = AutoZonedTest.Zoned.class)
+@DomainModel(annotatedClasses = DefaultZonedTest.Zoned.class)
 @SessionFactory
-@ServiceRegistry(settings = @Setting(name = AvailableSettings.TIMEZONE_DEFAULT_STORAGE, value = "AUTO"))
-public class AutoZonedTest {
+public class DefaultZonedTest {
 
 	@Test void test(SessionFactoryScope scope) {
 		ZonedDateTime nowZoned = ZonedDateTime.now().withZoneSameInstant( ZoneId.of("CET") );
@@ -45,8 +42,14 @@ public class AutoZonedTest {
 				assertEquals( nowZoned.toInstant(), z.zonedDateTime.toInstant() );
 				assertEquals( nowOffset.toInstant(), z.offsetDateTime.toInstant() );
 			}
-			assertEquals( nowZoned.toOffsetDateTime().getOffset(), z.zonedDateTime.toOffsetDateTime().getOffset() );
-			assertEquals( nowOffset.getOffset(), z.offsetDateTime.getOffset() );
+			if ( scope.getSessionFactory().getJdbcServices().getDialect().getTimeZoneSupport() == TimeZoneSupport.NATIVE ) {
+				assertEquals( nowZoned.toOffsetDateTime().getOffset(), z.zonedDateTime.toOffsetDateTime().getOffset() );
+				assertEquals( nowOffset.getOffset(), z.offsetDateTime.getOffset() );
+			}
+			else {
+				assertEquals( ZoneId.of("Z"), z.zonedDateTime.getZone() );
+				assertEquals( ZoneOffset.ofHours(0), z.offsetDateTime.getOffset() );
+			}
 		});
 	}
